moved location responder to service, added basic foreground notification and vibration for testing purposes

+++ app/src/main/java/com/iboism/gpxrecorder/model/TrackPoint.kt
@@ -11,9 +11,9 @@ import io.realm.annotations.PrimaryKey
 
 open class TrackPoint(
         @PrimaryKey var identifier: Long = UUIDHelper.random(),
-        var lat: Float = 0f,
-        var lon: Float = 0f,
-        var ele: Float = 0f,
+        var lat: Double = 0.0,
+        var lon: Double = 0.0,
+        var ele: Double = 0.0,
         var time: String = DateTimeFormatHelper.formatDate()
 ) : RealmObject(), XmlSerializable {
     override fun getXmlString(): String {
